Create styleForDashboard.css
this is the stylesheet used for designing the dashboard page that a logged-in user will see. It is not complete yet as more styling is required for the dashboard, this is just the initial version of it. Updated stylesheet will be uploaded when they are done.
(and they are not done yet 😄 )